weston: remove weston.service, its provided by other layers
meta-genivi-dev/meta-ivi does not follow the pattern set by poky
of having weston startup handled by weston-init.bb. This makes
integrating other layers, such as bsps, that do more difficult.
At the same time meta-genivi-dev is not accepting patches for
Genivi 10 so we can not change the behaviour even if we wanted
to. As a result simply remove the weston.service file provided
here, it is provided by other layers.